+++ etils/epy/__init__.py
@@ -18,9 +18,10 @@ import sys
 
 from etils.epy.backports import cached_property
 from etils.epy.env_utils import is_notebook
+from etils.epy.itertools import groupby
+from etils.epy.itertools import zip_dict
 from etils.epy.py_utils import issubclass_ as issubclass  # pylint: disable=redefined-builtin
 from etils.epy.py_utils import StrEnum
-from etils.epy.py_utils import zip_dict
 from etils.epy.reraise_utils import maybe_reraise
 from etils.epy.reraise_utils import reraise
 from etils.epy.text_utils import dedent